Fix headless mode
vkMapMemory failed because the memory lacked
VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT. Require that bit.

Drop the requirement for VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT because (a)
it is unneeded and (b) it may intefere with finding a type that supports
VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT.

Fail if protected mode is requested on the cmdline because host-visible
protected memory is useless and contradictory.

With these fixes, find_image_memory() is now independent of images. So
rename it choose_memory_type_index().